# дз_2/app/src/resources/openapi/marketplace.yaml
openapi: 3.0.3
info:
  title: Marketplace API
  version: 1.0.0
servers:
  - url: http://localhost:8000
paths:
  /auth/register:
    post:
      tags:
        - Auth
      summary: Register user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthRegisterRequest"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthRegisterResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
  /auth/login:
    post:
      tags:
        - Auth
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthLoginRequest"
      responses:
        "200":
          description: Tokens
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        "401":
          $ref: "#/components/responses/AuthError"
        "400":
          $ref: "#/components/responses/ValidationError"
  /auth/refresh:
    post:
      tags:
        - Auth
      summary: Refresh access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthRefreshRequest"
      responses:
        "200":
          description: New access token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        "401":
          $ref: "#/components/responses/RefreshError"
        "400":
          $ref: "#/components/responses/ValidationError"
  /products:
    get:
      tags:
        - Products
      summary: List products
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            minimum: 0
            default: 0
        - in: query
          name: size
          schema:
            type: integer
            minimum: 1
            default: 20
        - in: query
          name: status
          schema:
            $ref: "#/components/schemas/ProductStatus"
        - in: query
          name: category
          schema:
            type: string
      responses:
        "200":
          description: Product list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductListResponse"
        "401":
          $ref: "#/components/responses/AuthError"
    post:
      tags:
        - Products
      summary: Create product
      description: "Allowed roles: SELLER (own products), ADMIN (any)."
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductCreate"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/AuthError"
        "403":
          $ref: "#/components/responses/AccessDenied"
  /products/{id}:
    get:
      tags:
        - Products
      summary: Get product by ID
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductResponse"
        "401":
          $ref: "#/components/responses/AuthError"
        "404":
          $ref: "#/components/responses/ProductNotFound"
    put:
      tags:
        - Products
      summary: Update product
      description: "Allowed roles: SELLER (own products), ADMIN (any)."
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/AuthError"
        "403":
          $ref: "#/components/responses/AccessDenied"
        "404":
          $ref: "#/components/responses/ProductNotFound"
    delete:
      tags:
        - Products
      summary: Archive product
      description: "Allowed roles: SELLER (own products), ADMIN (any)."
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Archived
        "401":
          $ref: "#/components/responses/AuthError"
        "403":
          $ref: "#/components/responses/AccessDenied"
        "404":
          $ref: "#/components/responses/ProductNotFound"
  /orders:
    post:
      tags:
        - Orders
      summary: Create order
      description: "Allowed roles: USER (own orders), ADMIN (any)."
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderCreateRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/AuthError"
        "403":
          $ref: "#/components/responses/AccessDenied"
        "404":
          $ref: "#/components/responses/ProductNotFound"
        "409":
          $ref: "#/components/responses/OrderBusinessError"
        "422":
          $ref: "#/components/responses/PromoCodeError"
        "429":
          $ref: "#/components/responses/RateLimitError"
  /orders/{id}:
    get:
      tags:
        - Orders
      summary: Get order by ID
      description: "Allowed roles: USER (own orders), ADMIN (any)."
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderResponse"
        "401":
          $ref: "#/components/responses/AuthError"
        "403":
          $ref: "#/components/responses/AccessDenied"
        "404":
          $ref: "#/components/responses/OrderNotFound"
    put:
      tags:
        - Orders
      summary: Update order
      description: "Allowed roles: USER (own orders), ADMIN (any)."
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderUpdateRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/AuthError"
        "403":
          $ref: "#/components/responses/AccessDenied"
        "404":
          $ref: "#/components/responses/OrderNotFound"
        "409":
          $ref: "#/components/responses/OrderBusinessError"
        "422":
          $ref: "#/components/responses/PromoCodeError"
        "429":
          $ref: "#/components/responses/RateLimitError"
  /orders/{id}/cancel:
    post:
      tags:
        - Orders
      summary: Cancel order
      description: "Allowed roles: USER (own orders), ADMIN (any)."
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Canceled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderResponse"
        "401":
          $ref: "#/components/responses/AuthError"
        "403":
          $ref: "#/components/responses/AccessDenied"
        "404":
          $ref: "#/components/responses/OrderNotFound"
        "409":
          $ref: "#/components/responses/OrderBusinessError"
  /promo-codes:
    post:
      tags:
        - PromoCodes
      summary: Create promo code
      description: "Allowed roles: SELLER, ADMIN."
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PromoCodeCreate"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PromoCodeResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/AuthError"
        "403":
          $ref: "#/components/responses/AccessDenied"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            validation:
              value:
                error_code: VALIDATION_ERROR
                message: Validation error
                details:
                  fields:
                    - field: name
                      message: Must not be empty
    AuthError:
      description: Authentication error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    RefreshError:
      description: Refresh token error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    AccessDenied:
      description: Access denied
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ProductNotFound:
      description: Product not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    OrderBusinessError:
      description: Order business error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    PromoCodeError:
      description: Promo code error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    AuthRegisterRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
          maxLength: 255
        role:
          $ref: "#/components/schemas/UserRole"
      required:
        - email
        - password
    AuthRegisterResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        role:
          $ref: "#/components/schemas/UserRole"
        created_at:
          type: string
          format: date-time
      required:
        - id
        - email
        - role
        - created_at
    AuthLoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
          maxLength: 255
      required:
        - email
        - password
    AuthRefreshRequest:
      type: object
      properties:
        refresh_token:
          type: string
      required:
        - refresh_token
    TokenResponse:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        token_type:
          type: string
      required:
        - access_token
        - refresh_token
        - token_type
    ProductStatus:
      type: string
      enum: [ACTIVE, INACTIVE, ARCHIVED]
...
    ProductListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ProductResponse"
        totalElements:
          type: integer
        page:
          type: integer
        size:
          type: integer
      required:
        - items
        - totalElements
        - page
        - size
...